requirements.txt
# 核心依赖（兼容Python 3.6+）
pandas>=1.1.0,<2.0.0
openpyxl>=3.0.0
xlrd>=1.2.0

# 数据库驱动
SQLAlchemy>=1.4.0,<2.0.0
PyMySQL>=1.0.0
psycopg2-binary>=2.8.0
pyodbc>=4.0.0
oracledb>=1.0.0

# Web框架
Flask>=2.0.0,<3.0.0
Flask-CORS>=3.0.0

# 工具库
python-dotenv>=0.19.0
pyyaml>=5.4.0
jinja2>=3.0.0
colorama>=0.4.4
click>=8.0.0
schedule>=1.1.0
waitress>=2.0.0
